login screen UI Done
